+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Changes mentioned under 3.1.x version have not been published yet. Will be released when v3.1 is stable.
 
+## [3.1.16] - 2025-10-04
+- Updated SmartIdRestConnector to use v3.1 notification-based certificate choice endpoint
+- Added AccountUnusableException to handle ACCOUNT_UNUSABLE endResult from session status response
+
 ## [3.1.15] - 2025-09-17 
 - Added CallbackUrlUtil to generate callback URL with token and provides method to validate sessionSecretDigest
 - Updated DeviceLinkAuthenticationResponseValidator to also validate userChallenge and userChallengeVerifier same device flows.